Port phala-node-rpc-ext changes
