+++ crates/phala-node-rpc-ext/src/lib.rs
@@ -1,6 +1,7 @@
 use std::marker::PhantomData;
 use std::sync::Arc;
 
+use codec::Encode;
 use jsonrpc_derive::rpc;
 use mq_seq::Error as MqSeqError;
 use pallet_mq_runtime_api::MqApi;
@@ -33,6 +34,11 @@ pub trait NodeRpcExtApi<BlockHash> {
         from: BlockHash,
         to: BlockHash,
     ) -> Result<GetStorageChangesResponse, StorageChangesError>;
+
+    /// Get storage changes made by given block.
+    /// Returns `hex_encode(scale_encode(StorageChanges))`
+    #[rpc(name = "pha_getStorageChangesAt")]
+    fn get_storage_changes_at(&self, block: BlockHash) -> Result<String, StorageChangesError>;
 
     /// Return the next mq sequence number for given sender which take the ready transactions in count.
     #[rpc(name = "pha_getMqNextSequence")]
@@ -93,6 +99,13 @@ where
                 to,
             )
         }
+    }
+
+    fn get_storage_changes_at(&self, block: Block::Hash) -> Result<String, StorageChangesError> {
+        let changes = self.get_storage_changes(block, block)?;
+        // get_storage_changes never returns empty vec without error.
+        let encoded = changes[0].encode();
+        Ok(impl_serde::serialize::to_hex(&encoded, false))
     }
 
     fn get_mq_seq(&self, sender_hex: String) -> Result<u64, MqSeqError> {
