+++ rust-toolchain
@@ -1,4 +1,4 @@
 [toolchain]
-channel = "nightly-2021-09-01"
+channel = "nightly-2021-11-11"
 components = ["rustfmt"]
 targets = ["wasm32-unknown-unknown"]
